chore: move custom icons in a single file
