+++ src/application/sidebar/icons.tsx
@@ -1,0 +1,40 @@
+export const BotIcon = () => (
+  <svg className="self-center" height="24" viewBox="0 0 24 24" width="24" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M21.8971 9.82502V5.53971C21.8971 4.95394 21.6653 4.39217 21.2528 3.97798C20.8403 3.56378 20.2808 3.33109 19.6974 3.33109H13.0983V2.88224C13.4338 2.57966 13.6482 2.14456 13.6482 1.65646C13.6482 1.21714 13.4744 0.795813 13.165 0.485166C12.8556 0.174519 12.436 0 11.9985 0C11.5609 0 11.1413 0.174519 10.8319 0.485166C10.5225 0.795813 10.3487 1.21714 10.3487 1.65646C10.3487 2.14456 10.5632 2.57966 10.8986 2.88224V3.33109H4.29954C3.71615 3.33109 3.15664 3.56378 2.74412 3.97798C2.3316 4.39217 2.09985 4.95394 2.09985 5.53971V9.85042L2.02066 9.85594C1.74337 9.87578 1.48388 10.0004 1.29448 10.2047C1.10508 10.409 0.999854 10.6778 1 10.9569V13.1655C1 13.4584 1.11588 13.7393 1.32214 13.9464C1.5284 14.1535 1.80815 14.2699 2.09985 14.2699V19.7914C2.09985 20.3771 2.3316 20.9389 2.74412 21.3531C3.15664 21.7673 3.71615 22 4.29954 22H19.6974C20.2808 22 20.8403 21.7673 21.2528 21.3531C21.6653 20.9389 21.8971 20.3771 21.8971 19.7914V14.2699C22.1888 14.2699 22.4685 14.1535 22.6748 13.9464C22.8811 13.7393 22.9969 13.4584 22.9969 13.1655V11.0254C23.0097 10.854 22.9826 10.682 22.9177 10.5229C22.6956 9.98404 22.2193 9.85483 21.8971 9.82502ZM6.49923 9.95693C6.49923 8.73778 7.23833 7.74832 8.149 7.74832C9.05968 7.74832 9.79877 8.73778 9.79877 9.95693C9.79877 11.1761 9.05968 12.1655 8.149 12.1655C7.23833 12.1655 6.49923 11.1761 6.49923 9.95693ZM16.3957 17.5828C15.2947 17.5795 7.59908 17.5828 7.59908 17.5828V15.3742C7.59908 15.3742 15.2991 15.372 16.4001 15.3742L16.3957 17.5828ZM15.8479 12.1655C14.9373 12.1655 14.1982 11.1761 14.1982 9.95693C14.1982 8.73778 14.9373 7.74832 15.8479 7.74832C16.7586 7.74832 17.4977 8.73778 17.4977 9.95693C17.4977 11.1761 16.7586 12.1655 15.8479 12.1655Z"
+      fill="currentcolor"
+    />
+  </svg>
+);
+
+export const FilesIcon = () => (
+  <div className="flex h-6 w-6 shrink-0 content-center justify-center self-center">
+    <svg className="self-center" height="18" viewBox="0 0 20 18" width="20" xmlns="http://www.w3.org/2000/svg">
+      <path
+        clipRule="evenodd"
+        d="M17 18H3C1.874 18 1.074 17.509 0.588 16.834C0.210232 16.2966 0.00512901 15.6569 0 15V3C0 2.20435 0.316071 1.44129 0.87868 0.87868C1.44129 0.316071 2.20435 0 3 0H13C13.7956 0 14.5587 0.316071 15.1213 0.87868C15.6839 1.44129 16 2.20435 16 3V8H19C19.2652 8 19.5196 8.10536 19.7071 8.29289C19.8946 8.48043 20 8.73478 20 9V15C20 15.493 19.86 16.211 19.412 16.834C18.925 17.51 18.125 18 17 18ZM18 11C18 10.7348 17.8946 10.4804 17.7071 10.2929C17.5196 10.1054 17.2652 10 17 10C16.7348 10 16.4804 10.1054 16.2929 10.2929C16.1054 10.4804 16 10.7348 16 11V15C16 15.2652 16.1054 15.5196 16.2929 15.7071C16.4804 15.8946 16.7348 16 17 16C17.2652 16 17.5196 15.8946 17.7071 15.7071C17.8946 15.5196 18 15.2652 18 15V11Z"
+        fill="currentcolor"
+      />
+    </svg>
+  </div>
+);
+
+export const PenIcon = () => (
+  <svg className="self-center" height="24" viewBox="0 0 24 24" width="24" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M2 4.621C2.00002 4.52206 2.02939 4.42536 2.0844 4.34312C2.1394 4.26089 2.21757 4.19682 2.309 4.15903C2.40043 4.12123 2.50102 4.11141 2.59804 4.13081C2.69505 4.15021 2.78413 4.19795 2.854 4.268L8.864 10.278C8.99 10.404 9.034 10.588 9.014 10.765C8.96513 11.1778 9.04609 11.5955 9.24561 11.9602C9.44513 12.3248 9.7533 12.6182 10.1273 12.7996C10.5013 12.981 10.9225 13.0414 11.3324 12.9723C11.7422 12.9033 12.1204 12.7083 12.4143 12.4143C12.7083 12.1204 12.9033 11.7422 12.9723 11.3324C13.0414 10.9225 12.981 10.5013 12.7996 10.1273C12.6182 9.7533 12.3248 9.44513 11.9602 9.24561C11.5955 9.04609 11.1778 8.96513 10.765 9.014C10.6778 9.02727 10.5887 9.02068 10.5043 8.99471C10.42 8.96873 10.3426 8.92405 10.278 8.864L4.268 2.854C4.19801 2.78419 4.15029 2.6952 4.13086 2.59828C4.11143 2.50136 4.12116 2.40085 4.15883 2.30946C4.1965 2.21807 4.26042 2.13989 4.3425 2.08481C4.42458 2.02973 4.52115 2.00022 4.62 2H11C12.4448 1.99968 13.8685 2.34722 15.1507 3.01324C16.4329 3.67927 17.5359 4.6442 18.3664 5.82646C19.197 7.00873 19.7307 8.37358 19.9224 9.80564C20.1141 11.2377 19.9583 12.6949 19.468 14.054L21.708 16.293C21.8955 16.4805 22.0008 16.7348 22.0008 17C22.0008 17.2652 21.8955 17.5195 21.708 17.707L17.708 21.707C17.6151 21.8 17.5048 21.8737 17.3834 21.9241C17.262 21.9744 17.1319 22.0003 17.0005 22.0003C16.8691 22.0003 16.739 21.9744 16.6176 21.9241C16.4962 21.8737 16.3859 21.8 16.293 21.707L14.054 19.468C12.6949 19.9583 11.2377 20.1141 9.80564 19.9224C8.37358 19.7307 7.00873 19.197 5.82646 18.3664C4.6442 17.5359 3.67927 16.4329 3.01324 15.1507C2.34722 13.8685 1.99968 12.4448 2 11V4.621Z"
+      fill="currentcolor"
+    />
+  </svg>
+);
+
+export const ReferencesIcon = () => (
+  <div className="flex h-6 w-6 shrink-0 content-center justify-center self-center">
+    <svg className="self-center" height="20" viewBox="0 0 16 20" width="16" xmlns="http://www.w3.org/2000/svg">
+      <path
+        d="M4 4H3C2.73478 4 2.48043 3.89464 2.29289 3.70711C2.10536 3.51957 2 3.26522 2 3C2 2.73478 2.10536 2.48043 2.29289 2.29289C2.48043 2.10536 2.73478 2 3 2H14V1C14 0.734784 13.8946 0.48043 13.7071 0.292893C13.5196 0.105357 13.2652 0 13 0L2 0C1.46957 0 0.960859 0.210714 0.585786 0.585786C0.210714 0.960859 0 1.46957 0 2V18C0 19.1 0.9 20 2 20H14C14.5304 20 15.0391 19.7893 15.4142 19.4142C15.7893 19.0391 16 18.5304 16 18V5C16 4.73478 15.8946 4.48043 15.7071 4.29289C15.5196 4.10536 15.2652 4 15 4H8V12L6 10L4 12V4Z"
+        fill="currentcolor"
+      />
+    </svg>
+  </div>
+);
