--- src/application/sidebar/icons/PenIcon.tsx
@@ -1,8 +1,0 @@
-export const PenIcon = () => (
-  <svg className="self-center" height="24" viewBox="0 0 24 24" width="24" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M2 4.621C2.00002 4.52206 2.02939 4.42536 2.0844 4.34312C2.1394 4.26089 2.21757 4.19682 2.309 4.15903C2.40043 4.12123 2.50102 4.11141 2.59804 4.13081C2.69505 4.15021 2.78413 4.19795 2.854 4.268L8.864 10.278C8.99 10.404 9.034 10.588 9.014 10.765C8.96513 11.1778 9.04609 11.5955 9.24561 11.9602C9.44513 12.3248 9.7533 12.6182 10.1273 12.7996C10.5013 12.981 10.9225 13.0414 11.3324 12.9723C11.7422 12.9033 12.1204 12.7083 12.4143 12.4143C12.7083 12.1204 12.9033 11.7422 12.9723 11.3324C13.0414 10.9225 12.981 10.5013 12.7996 10.1273C12.6182 9.7533 12.3248 9.44513 11.9602 9.24561C11.5955 9.04609 11.1778 8.96513 10.765 9.014C10.6778 9.02727 10.5887 9.02068 10.5043 8.99471C10.42 8.96873 10.3426 8.92405 10.278 8.864L4.268 2.854C4.19801 2.78419 4.15029 2.6952 4.13086 2.59828C4.11143 2.50136 4.12116 2.40085 4.15883 2.30946C4.1965 2.21807 4.26042 2.13989 4.3425 2.08481C4.42458 2.02973 4.52115 2.00022 4.62 2H11C12.4448 1.99968 13.8685 2.34722 15.1507 3.01324C16.4329 3.67927 17.5359 4.6442 18.3664 5.82646C19.197 7.00873 19.7307 8.37358 19.9224 9.80564C20.1141 11.2377 19.9583 12.6949 19.468 14.054L21.708 16.293C21.8955 16.4805 22.0008 16.7348 22.0008 17C22.0008 17.2652 21.8955 17.5195 21.708 17.707L17.708 21.707C17.6151 21.8 17.5048 21.8737 17.3834 21.9241C17.262 21.9744 17.1319 22.0003 17.0005 22.0003C16.8691 22.0003 16.739 21.9744 16.6176 21.9241C16.4962 21.8737 16.3859 21.8 16.293 21.707L14.054 19.468C12.6949 19.9583 11.2377 20.1141 9.80564 19.9224C8.37358 19.7307 7.00873 19.197 5.82646 18.3664C4.6442 17.5359 3.67927 16.4329 3.01324 15.1507C2.34722 13.8685 1.99968 12.4448 2 11V4.621Z"
-      fill="currentcolor"
-    />
-  </svg>
-);
